move files in src to root so package will contain all necessary files
