+++ lib/interfaces.ts
